Add filters & cleanup codebase & refactor VacanciesList
