--- src/blocks/EventList/EventList.css
@@ -1,9 +1,0 @@
-.event-list-filter {
-  border-block-end: 1px solid var(--color-black);
-}
-
-.empty-message {
-  padding: var(--spacing-104) var(--spacing-48);
-  text-align: center;
-  margin: 0;
-}
