+++ src/blocks/HighlightedVacancies/HightlightedVacancies.astro
@@ -1,9 +1,9 @@
 ---
-import { VacancyDataList } from '@components/VacancyDataList';
+import { VacancyList } from '@components/VacancyList';
 import type { VacancyListItemFragment } from '@lib/types/datocms';
 import { getRandomItems } from '@lib/content-collections';
 
 const vacancies = await getRandomItems<VacancyListItemFragment>('vacancies', 6);
 ---
 
-<VacancyDataList vacancies={vacancies} />
+<VacancyList vacancies={vacancies} />
