--- src/blocks/MemberList/MemberList.css
@@ -1,12 +1,0 @@
-.member-list {
-  list-style-type: "";
-  padding: 0px;
-  margin: 0px;
-}
-
-.empty-message {
-  padding: var(--spacing-104) var(--spacing-48);
-  text-align: center;
-  border-block-start: 1px solid var(--color-black);
-  margin: 0;
-}
