--- src/blocks/MemberList/_MemberList.query.graphql
@@ -1,6 +1,0 @@
-query MemberList {
-  provinces: allProvinces {
-    id
-    name
-  }
-}
