--- src/blocks/MemberList/index.ts
@@ -1,1 +1,0 @@
-export { default as MemberList } from './MemberList.astro';
