--- src/blocks/PublicationList/PublicationList.astro
@@ -1,24 +1,0 @@
----
-import { getDehydratedState } from '@lib/react-query';
-import {
-  PublicationList as PublicationListComponent,
-  loader,
-} from './PublicationList.tsx';
-
-const searchParams = Astro.url.searchParams;
-const initialParams = {
-  page: searchParams.get('page') ?? '1',
-};
-const initialData = await loader(initialParams);
-const initialUrl = Astro.url.href;
----
-
-<PublicationListComponent
-  client:load
-  initialParams={initialParams}
-  initialState={getDehydratedState(
-    ['publications', initialParams],
-    initialData
-  )}
-  initialUrl={initialUrl}
-/>
