--- src/blocks/PublicationList/PublicationList.css
@@ -1,5 +1,0 @@
-.empty-message {
-  padding: var(--spacing-104) var(--spacing-48);
-  text-align: center;
-  margin: 0;
-}
