--- src/blocks/PublicationList/PublicationList.query.graphql
@@ -1,41 +1,0 @@
-query PublicationsList(
-  $first: IntType!
-  $skip: IntType!
-  $filter: PublicationModelFilter
-) {
-  publicationsMeta: _allPublicationsMeta(filter: $filter) {
-    count
-  }
-  tags: allSubjects {
-    id
-    name
-  }
-  publications: allPublications(
-    first: $first
-    skip: $skip
-    filter: $filter
-    orderBy: _createdAt_DESC
-  ) {
-    id
-    title
-    slug
-    _createdAt
-    tags {
-      name
-    }
-    banner {
-      alt
-      responsiveImage {
-        src
-        srcSet
-        width
-        height
-        alt
-        title
-        base64
-        bgColor
-        sizes
-      }
-    }
-  }
-}
