--- src/blocks/VacancyList/VacancyList.astro
@@ -1,37 +1,0 @@
----
-import { getDehydratedState } from '@lib/react-query';
-import { getCollection } from 'astro:content';
-import { VacancyList as VacancyListComponent, loader } from './VacancyList';
-
-const searchParams = Astro.url.searchParams;
-const initialParams = {
-  page: searchParams.get('page') ?? '1',
-  zoek: searchParams.get('zoek') ?? '',
-  voertaal: searchParams.get('voertaal') ?? '',
-  soort: searchParams.get('soort') ?? '',
-  uren: searchParams.get('uren') ?? '',
-  provincie: searchParams.get('provincie') ?? '',
-};
-const initialUrl = Astro.url.href;
-const initialData = await loader(initialParams);
-
-// these filter values are aggregated from all vacancies,
-// since they do not exist as their own collection in the CMS
-const vacancyEmploymentTypes = await getCollection('vacancyEmploymentTypes');
-const vacancyHours = await getCollection('vacancyHours');
-const vacancyLanguages = await getCollection('vacancyLanguages');
-
-const filterValues = {
-  employmentTypes: vacancyEmploymentTypes,
-  hours: vacancyHours,
-  languages: vacancyLanguages,
-};
----
-
-<VacancyListComponent
-  client:load
-  initialParams={initialParams}
-  initialState={getDehydratedState(['vacancies', initialParams], initialData)}
-  initialUrl={initialUrl}
-  filterValues={filterValues}
-/>
