--- src/blocks/VacancyList/VacancyList.query.graphql
@@ -1,18 +1,0 @@
-#import '@blocks/VacancyList/VacancyListItem.fragment.graphql'
-
-query VacancyList(
-  $first: IntType!
-  $skip: IntType!
-  $filter: VacancyModelFilter
-) {
-  provinces: allProvinces {
-    id
-    name
-  }
-  vacanciesMeta: _allVacanciesMeta(filter: $filter) {
-    count
-  }
-  vacancies: allVacancies(first: $first, skip: $skip, filter: $filter) {
-    ...VacancyListItem
-  }
-}
