--- src/components/NewPagination/Pagination.css
@@ -1,35 +1,0 @@
-.pagination {
-  display: flex;
-  justify-content: center;
-  padding: var(--spacing-48);
-  gap: 1.5rem;
-  border-top: 1px solid var(--color-black);
-}
-
-.pagination-list {
-  display: flex;
-  
-}
-
-.pagination-list ul {
-  list-style-type: "";
-  display: flex;
-  margin: 0px;
-  padding: 0px;
-}
-
-.pagination__ellipsis {
-  display: inline-flex;
-  color: var(--black);
-  align-items: center;
-  justify-content: center;
-  padding-inline: 0.5rem;
-}
-
-.pagination-list__item {
-  padding-inline: 0.5rem;
-}
-
-.pagination-list__item[aria-current="page"] {
-  text-decoration: underline;
-}
