+++ src/components/Pagination/Pagination.css
@@ -1,19 +1,21 @@
 .pagination {
   display: flex;
-  gap: 1rem;
   justify-content: center;
+  padding: var(--spacing-48);
+  gap: 1.5rem;
+  border-top: 1px solid var(--color-black);
 }
 
-.pagination__pages {
+.pagination-list {
   display: flex;
+  
 }
 
-.pagination__page {
-  padding-inline: 0.5rem;
-}
-
-.pagination__page--active {
-  text-decoration: underline;
+.pagination-list ul {
+  list-style-type: "";
+  display: flex;
+  margin: 0px;
+  padding: 0px;
 }
 
 .pagination__ellipsis {
@@ -22,4 +24,12 @@
   align-items: center;
   justify-content: center;
   padding-inline: 0.5rem;
+}
+
+.pagination-list__item {
+  padding-inline: 0.5rem;
+}
+
+.pagination-list__item[aria-current="page"] {
+  text-decoration: underline;
 }
