--- src/components/Pagination/index.ts
@@ -1,1 +1,0 @@
-export { Pagination } from './Pagination';
