+++ src/components/VacancyList/VacancyList.css
@@ -1,0 +1,30 @@
+.vacancy-list {
+  box-shadow:
+    0 1px 0 0 var(--color-black),
+    0 -1px 0 0 var(--color-black);
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacing-40);
+  list-style-type: none;
+  margin: 0;
+  padding: var(--spacing-40) var(--spacing-16);
+}
+
+.vacancy-list:empty {
+  padding: 0;
+}
+
+@media screen and (min-width: 481px) {
+  .vacancy-list {
+    padding: var(--spacing-40) var(--spacing-48);
+  }
+}
+
+@media screen and (min-width: 1225px) {
+  .vacancy-list {
+    align-items: center;
+    display: grid;
+    row-gap: var(--spacing-32);
+    grid-template-columns: 1fr repeat(5, auto);
+  }
+}
