+++ src/components/VacancyListItem/VacancyListItem.fragment.graphql
